TEMP

+++ kernel/includes/mm/dmas.h
@@ -5,8 +5,8 @@
 #include <task/constant.h>
 #include <mm/map.h>
 
-#define mm_dmas_phys2Virt(physAddr) ((void *)((physAddr) + mm_dmas_addrSt))
-#define mm_dmas_virt2Phys(virtAddr) ((u64)((virtAddr) - mm_dmas_addrSt))
+#define mm_dmas_phys2Virt(physAddr) ((void *)((u64)(physAddr) + mm_dmas_addrSt))
+#define mm_dmas_virt2Phys(virtAddr) ((u64)((u64)(virtAddr) - mm_dmas_addrSt))
 
 int mm_dmas_init();
